+++ savvy-bindgen/src/parse/savvy_enum.rs
@@ -1,4 +1,4 @@
-use syn::parse_quote;
+use syn::{parse_quote, ItemEnum, ItemImpl};
 
 use crate::extract_docs;
 
@@ -40,5 +40,65 @@ impl SavvyEnum {
             ty,
             variants,
         })
+    }
+
+    pub fn generate_enum_with_discriminant(&self) -> ItemEnum {
+        let ty = &self.ty;
+        let attrs = &self.attrs;
+
+        let variants_tweaked = self
+            .variants
+            .iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let lit_i = syn::LitInt::new(&i.to_string(), v.span());
+                parse_quote!(#v = #lit_i)
+            })
+            .collect::<Vec<syn::Variant>>();
+
+        parse_quote!(
+            #(#attrs)*
+            pub enum #ty {
+                #(#variants_tweaked),*
+            }
+        )
+    }
+
+    pub fn generate_try_from_impls(&self) -> Vec<ItemImpl> {
+        let ty = &self.ty;
+
+        let match_arms = self
+            .variants
+            .iter()
+            .enumerate()
+            .map(|(i, v)| {
+                let lit_i = syn::LitInt::new(&i.to_string(), v.span());
+                parse_quote!(#lit_i => Ok(#ty::#v))
+            })
+            .collect::<Vec<syn::Arm>>();
+
+        vec![
+            parse_quote!(
+            impl TryFrom<#ty> for savvy::Sexp {
+                type Error = savvy::Error;
+
+                fn try_from(value: #ty) -> savvy::Result<Self> {
+                    (value as i32).try_into()
+                }
+            }),
+            parse_quote!(
+                impl TryFrom<savvy::Sexp> for #ty {
+                    type Error = savvy::Error;
+
+                    fn try_from(value: savvy::Sexp) -> savvy::Result<Self> {
+                        let i = <i32>::try_from(value)?;
+                        match i {
+                            #(#match_arms),*,
+                            _ => Err("Unexpected enum variant".into()),
+                        }
+                    }
+                }
+            ),
+        ]
     }
 }
